requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.3
django-appconf==1.0.6
django-ckeditor==6.7.1
django-crispy-forms==2.1
django-embed-video==1.4.9
django-imagekit==5.0.0
django-import-export==3.3.8
django-js-asset==2.2.0
et-xmlfile==1.1.0
idna==3.6
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
pilkit==3.0
pillow==10.3.0
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.10.0
urllib3==2.2.1
xlrd==2.0.1
xlwt==1.3.0
